Creat index and js file
I did a lot here and forgot to commit. This creates index.html and
rps.js, and links rps.js to index.html. It creats a function for
randomly choosing rock paper or scissors for the computer. It creats
includes a previously created function to capitalize the first letter of
a string input. It also creates a function to play a round of the rps
that takes in player and computer selection and returns a message based
on the outcome. Finally, includes some js to play a basic game where the
payer alwasy chooses rock and the console logs teh winner, as a test.