Create for loop in game()
Added code for the for loop contained in game(). Creates four vars, to store player and computer wins as well as strings to display score. Establishes for loop as well as code within to take player selection via prompt, computer selection via computerPlay, and runs them through playRound. The results of playRound are stored in the var won, and this is used in an if/else to log a message to the console depending on who wins the round and to update and display the score string.