Add conditions after for loop in game()
Adds the win conditions after the for loop executes in game(). Once the for loop completes, the final score is stored as a string in the finalScore var. Then, an if/else tree compares playerWins to computerWins and logs a message to the console based on if the player wins, computer wins, or if there is a draw.