Misc fixes
Removes break statements form switch expression in computerPlay. Changes compSel to a var instead of a constant. Creates a var pSelFix that stores the properly formatted playerSelection for returning later in the score and finalScore strings. Game currently runs correctly unless the player doesn't enter an input, at which point it crashes. Needs to be updated to handle no input as well as incorrect inputs.