Flesh out game()
Creates constants for each gameplay button. Adds event listeners to each button in game() that runs updateWinner on click, using playRound as the arg. playRound uses one of the choose functions to return the button chosen as playerSelection, and computerPlay is used for computerSelection. Once a button has been chosen, the loop runs updateGameOver. May change so that updateGameOver is ran as part of updateWinner so as to make sure the loop executes properly.